+++ src/constants.ts
@@ -1,0 +1,34 @@
+import path from "path";
+
+export const {
+  SPACE_ID,
+  MANAGEMENT_API_KEY,
+  GITHUB_WORKSPACE,
+  INPUT_MIGRATIONS_DIR,
+  INPUT_DELETE_FEATURE,
+  INPUT_SET_ALIAS,
+  INPUT_FEATURE_PATTERN,
+  INPUT_MASTER_PATTERN,
+  INPUT_VERSION_CONTENT_TYPE,
+  INPUT_VERSION_FIELD,
+} = process.env;
+
+export const DEFAULT_MIGRATIONS_DIR = "migrations";
+export const DEFAULT_MASTER_PATTERN = "master-[YYYY]-[MM]-[DD]-[mmss]";
+export const DEFAULT_FEATURE_PATTERN = "GH-[branch]";
+export const DEFAULT_VERSION_CONTENT_TYPE = "versionTracking";
+export const DEFAULT_VERSION_FIELD = "version";
+
+export const VERSION_CONTENT_TYPE =
+  INPUT_VERSION_CONTENT_TYPE || DEFAULT_VERSION_CONTENT_TYPE;
+export const FEATURE_PATTERN = INPUT_FEATURE_PATTERN || DEFAULT_FEATURE_PATTERN;
+export const MASTER_PATTERN = INPUT_MASTER_PATTERN || DEFAULT_MASTER_PATTERN;
+export const VERSION_FIELD = INPUT_VERSION_FIELD || DEFAULT_VERSION_FIELD;
+export const MIGRATIONS_DIR = path.join(
+  GITHUB_WORKSPACE,
+  INPUT_MIGRATIONS_DIR || DEFAULT_MIGRATIONS_DIR
+);
+
+export const CONTENTFUL_MASTER = "master";
+export const DELAY = 3000;
+export const MAX_NUMBER_OF_TRIES = 10;
